Split controllers/devices.py into a subpackage
Per the follow-up to #123 — same shape as the firmware-controller
split, applied to the now-2094-line ``controllers/devices.py``
which mixes device CRUD, file scanning, mDNS / MQTT bridging, and
the WS handler glue.

Subpackage layout:

- ``controllers/devices/__init__.py`` — re-exports
  ``DevicesController`` so existing
  ``from .controllers.devices import DevicesController`` keeps
  resolving.
- ``controllers/devices/constants.py`` — ANSI-conceal regex used
  by the secret redactor.
- ``controllers/devices/helpers.py`` — pure free functions
  (``_wipe_device_build_dir``, ``_remove_device_sidecars``,
  ``_validate_archive_configuration``, ``_redact_concealed_secrets``,
  ``_apply_featured_presets``, ``_normalize_pin_value``,
  ``_build_address_cache_args``) plus the ``friendly_name_slugify``
  re-export.
- ``controllers/devices/controller.py`` — ``DevicesController``
  itself + the scan / state-monitor / MQTT bridge.

No behaviour change. ``device_builder.py`` and the rest of the
package only ever imported ``DevicesController`` from the package
root; that path is unchanged via ``__init__.py`` re-export.

Tests mirror the production split:
``tests/test_{archive,create,delete,rename,import,
metadata_resolver,stream_cancel,ignored_devices_load,
friendly_name_slugify_import}_device*.py`` moved to
``tests/controllers/devices/test_*.py`` (filename prefix dropped
where redundant). Test files that primarily exercise sibling
modules (``_device_state_monitor``, ``_device_scanner``,
``_dns_cache``) stay at the top level since they're not the
devices-controller subject.

Two private-symbol monkeypatch families needed updates:

- ``ext_storage_path`` is imported into both ``controller.py``
  and ``helpers.py``; archive/delete fixtures now patch both
  bindings since ``_archive_single`` / ``_delete_single`` (in
  ``controller``) delegate to ``_wipe_device_build_dir`` /
  ``_remove_device_sidecars`` (in ``helpers``) and rebinding
  only one leaves the other path running against the real
  ``CORE.config_path``.
- ``devices_module`` aliases (``test_create``, ``test_import``,
  the ``remove_device_metadata`` exception case) repointed at
  ``controllers.devices.controller`` / ``controllers.devices.helpers``
  as appropriate.

701 tests pass; pre-commit clean.
