+++ tests/controllers/devices/__init__.py
@@ -1,0 +1,1 @@
+"""Tests for ``controllers/devices/`` — split mirrors the production subpackage."""
