+++ tests/controllers/firmware/test_refresh.py
@@ -255,11 +255,11 @@ async def test_refresh_after_compile_persists_hash_and_reloads(
         return "1a2b3c4d"
 
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.devices.set_device_metadata",
+        "esphome_device_builder.controllers.devices.controller.set_device_metadata",
         _fake_set_metadata,
     )
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.devices.compute_yaml_config_hash",
+        "esphome_device_builder.controllers.devices.controller.compute_yaml_config_hash",
         _fake_compute,
     )
 
@@ -292,11 +292,11 @@ async def test_refresh_after_compile_skips_persist_on_hash_failure(
         return None  # YAML didn't validate, subprocess failed, etc.
 
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.devices.set_device_metadata",
+        "esphome_device_builder.controllers.devices.controller.set_device_metadata",
         _fake_set_metadata,
     )
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.devices.compute_yaml_config_hash",
+        "esphome_device_builder.controllers.devices.controller.compute_yaml_config_hash",
         _fake_compute,
     )
 
@@ -325,7 +325,7 @@ async def test_refresh_after_upload_skips_hash_compute(tmp_path: Path, monkeypat
         return "deadbeef"
 
     monkeypatch.setattr(
-        "esphome_device_builder.controllers.devices.compute_yaml_config_hash",
+        "esphome_device_builder.controllers.devices.controller.compute_yaml_config_hash",
         _fake_compute,
     )
